refrectaor de la classe dans le model
